the dashboard refactored
