+++ src/Shared/common/styles/styles.css
@@ -1,0 +1,11 @@
+.avatar-container {
+  position: relative;
+  background-color: #fff;
+}
+.avatar {
+  position: absolute;
+  top: -64px;
+  display: table;
+  left: 45%;
+  transform: (-45%, 0);
+}
